requirements.txt
django<4
djangorestframework
uwsgi
django-rest-knox
django-cors-headers
dj-rest-auth[with_social]
requests
bs4